services/converter_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "converter_service"
version = "0.1.0"
description = "Lightweight CAD converter service"
authors = [{name="Michel Kruse"}]
dependencies = [
    "fastapi",
    "uvicorn",
    "open3d",
    "cascadio",
    "requests",
    "trimesh>=4.8.2",
    "python-multipart>=0.0.20",
    "gmsh>=4.11.0",
    "meshio>=5.3.0",
    "legacy-quadpy",
    "scikit-learn>=1.3.0",
    "stl-to-voxel",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
]



[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
